fix: Resolve generator disappearing after save and improve UI
- Fix toggle switch visibility for required field (was invisible when unchecked)
- Add save icon in header during edit mode, hide open icon to prevent confusion
- Fix API service returning wrong data structure causing generator to appear "not found"
