+++ gruenerator_frontend/src/assets/styles/components/custom-generator/field-editor-assistant.css
@@ -173,30 +173,31 @@
 .toggle-switch {
     position: relative;
     display: inline-block;
-    width: 50px; 
-    height: 28px; 
-    background-color: var(--border-subtle, rgba(255, 255, 255, 0.2)); 
-    border-radius: 14px; 
+    width: 50px;
+    height: 28px;
+    background-color: var(--dimmed, rgba(255, 255, 255, 0.15));
+    border-radius: 14px;
     cursor: pointer;
-    border: none;
+    border: 2px solid var(--border-subtle, rgba(255, 255, 255, 0.4));
     padding: 0;
-    transition: background-color 0.3s ease;
+    transition: background-color 0.3s ease, border-color 0.3s ease;
 }
 
 .toggle-switch-thumb {
     position: absolute;
-    top: 3px;
-    left: 3px;
-    width: 22px; 
+    top: 1px;
+    left: 1px;
+    width: 22px;
     height: 22px;
     background-color: var(--weiß, #ffffff);
-    border-radius: 50%; 
+    border-radius: 50%;
     transition: transform 0.3s ease;
 }
 
 
 .toggle-switch.checked {
-    background-color: var(--grashalm, #8ABD24); 
+    background-color: var(--grashalm, #8ABD24);
+    border-color: var(--grashalm, #8ABD24);
 }
 
 .toggle-switch.checked .toggle-switch-thumb {
