--- gruenerator_frontend/src/assets/styles/features/auth/profile-bubbles.css
@@ -1,47 +1,0 @@
-/* 
- * Bubble Animation Styles für den Labor-Tab
- */
-
-.bubble-container {
-  position: absolute;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  height: 10px;
-  overflow: visible;
-  pointer-events: none;
-  z-index: 1;
-}
-
-.bubble {
-  position: absolute;
-  border-radius: 50%;
-  transform-origin: center center;
-  background: radial-gradient(circle at 25% 25%, rgba(255, 255, 255, 0.5), rgba(161, 224, 186, 0.6));
-  box-shadow: 0 0 2px rgba(255, 255, 255, 0.8) inset;
-}
-
-/* Wrapper für Positionierung der Bubbles relativ zum Tab */
-.bubble-tab-wrapper {
-  position: relative;
-  display: inline-block;
-}
-
-/* Container für Positionierung */
-.bubbles-position-wrapper {
-  position: absolute;
-  bottom: 100%;
-  left: 0;
-  right: 0;
-  height: 0;
-  overflow: visible;
-  pointer-events: none;
-}
-
-/* Mobile Optimierung */
-@media (max-width: 767px) {
-  .bubble {
-    /* Kleinere Blasen auf mobilen Geräten */
-    transform: scale(0.8);
-  }
-}
